app/backend/requirements.txt
Flask==3.1.0 
flask-cors==5.0.1 
tensorflow==2.19.0 
torch==2.7.0 
torchvision==0.22.0 
numpy==2.0.0 
Pillow==11.2.1 
Werkzeug==3.1.3
